+++ src/Yaml.Localizer/YamlLocalizer.cs
@@ -9,8 +9,9 @@ namespace Yaml.Localizer
     /// </summary>
     public class YamlLocalizer
     {
-        private readonly CultureInfo defaultCulture = CultureInfo.CurrentCulture;
+        private readonly CultureInfo defaultCulture;
         private readonly List<MessageTranslations> translationMappings;
+        private CultureInfo currentCulture;
 
         /// <summary>
         /// Initializes a new instance of the <see cref="YamlLocalizer"/> class.
@@ -26,12 +27,9 @@ namespace Yaml.Localizer
             this.translationMappings = deserializer.Deserialize<List<MessageTranslations>>(input);
             this.defaultCulture = this.translationMappings.FirstOrDefault()?.Messages.Keys.FirstOrDefault()
                 ?? CultureInfo.CurrentCulture;
-        }
 
-        /// <summary>
-        /// Gets or sets the current culture for localization.
-        /// </summary>
-        public CultureInfo CurrentCulture { get; set;} = CultureInfo.CurrentCulture;
+            this.currentCulture = defaultCulture;
+        }
 
         /// <summary>
         /// Gets the localized string for the specified message ID.
@@ -40,19 +38,53 @@ namespace Yaml.Localizer
         /// <returns>The localized string.</returns>
         public string this[string id] => this.GetTranslation(id);
 
+        /// <summary>
+        /// Gets or sets the current culture for localization.
+        /// </summary>
+        [Obsolete("This propery will be removed in future releases. Please use the 'UseCultureOrDefault' method instead.")]
+        public CultureInfo CurrentCulture
+        {
+            get
+            {
+                return this.currentCulture;
+            }
+            set
+            {
+                this.currentCulture = value;
+            }
+        }
+
+        /// <summary>
+        /// Sets the specified valid locale; otherwise, the default value will be used.
+        /// </summary>
+        /// <param name="lang">The ISO code for the locale.</param>
+        public void UseCultureOrDefault(string? lang)
+        {
+            try
+            {
+                this.currentCulture = string.IsNullOrEmpty(lang)
+                    ? this.defaultCulture
+                    : new CultureInfo(lang);
+            }
+            catch
+            {
+                this.currentCulture = this.defaultCulture;
+            }
+        }
+
         private string GetTranslation(string id)
         {   
             var data = this.translationMappings.FirstOrDefault(t => t.Id == id)
                 ?? throw new KeyNotFoundException($"Translation ID '{id}' not found.");
 
             var result = data.Messages.FirstOrDefault(c =>
-                c.Key.TextInfo.CultureName == this.CurrentCulture.TextInfo.CultureName
+                c.Key.TextInfo.CultureName == this.currentCulture.TextInfo.CultureName
             ).Value ?? data.Messages.FirstOrDefault(c =>
-                c.Key.TwoLetterISOLanguageName == this.CurrentCulture.TwoLetterISOLanguageName
+                c.Key.TwoLetterISOLanguageName == this.currentCulture.TwoLetterISOLanguageName
             ).Value ?? data.Messages.FirstOrDefault(c => c.Key == this.defaultCulture).Value;
             
             return result ?? throw new KeyNotFoundException(
-                $"Translation for language '{this.CurrentCulture}' or default culture '{this.defaultCulture}' not found.");
+                $"Translation for language '{this.currentCulture}' or default culture '{this.defaultCulture}' not found.");
         }
     }
 }