+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Yaml.Localizer
 
+## Release 10.2.0
+
+- New method `UseCultureOrDefault`.
+- The `CurrentCulture` property is deprecated.
+
 ## Release 10.1.1
 
 - New `CurrentCulture` property to be used as the selected culture.